I added 3 extention method to organize the program.cs class as it was messy.
