+++ EndpointX/Extensions/EFCoreExtensions.cs
@@ -1,0 +1,18 @@
+﻿using EndpointX.Models.Data;
+using EndpointX.Models;
+using Microsoft.EntityFrameworkCore;
+
+namespace EndpointX.Extensions
+{
+    public static class EFCoreExtensions
+    {
+        public static IServiceCollection InjectDbContext(this IServiceCollection services, IConfiguration config)
+        {
+            var connectionString = config.GetConnectionString("DefaultConnection");
+            services.AddDbContext<ApplicationDbContext>(options =>
+                options.UseSqlServer(connectionString));
+
+            return services;
+        }
+    }
+}
