+++ EndpointX/Extensions/IdentityExtensions.cs
@@ -1,0 +1,65 @@
+﻿using EndpointX.Models.Data;
+using EndpointX.Models;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.IdentityModel.Tokens;
+using System.Text;
+
+namespace EndpointX.Extensions
+{
+    public static class IdentityExtensions
+    {
+        public static IServiceCollection AddIdentityHandlersAndStores(this IServiceCollection services)
+        {
+            services.AddAuthorization()
+                    .AddIdentityApiEndpoints<ApplicationUser>()
+                    .AddEntityFrameworkStores<ApplicationDbContext>();
+
+            return services;
+        }
+
+        public static IServiceCollection ConfigureIdentityOption(this IServiceCollection services)
+        {
+            services.Configure<IdentityOptions>(options =>
+            {
+                options.Password.RequireDigit = false;
+                options.Password.RequireUppercase = false;
+                options.Password.RequireLowercase = false;
+                options.User.RequireUniqueEmail = true;
+            });
+
+            return services;
+        }
+
+        //Auth = Authentication + Authorization
+        public static IServiceCollection AddIdentityAuth(this IServiceCollection services, IConfiguration config)
+        {
+            services.AddAuthentication(x =>
+            {
+                x.DefaultAuthenticateScheme =
+                x.DefaultChallengeScheme =
+                x.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
+            }).AddJwtBearer(y =>
+            {
+                y.SaveToken = false;
+                y.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuerSigningKey = true,
+                    IssuerSigningKey = new SymmetricSecurityKey(
+                            Encoding.UTF8.GetBytes(
+                                config["AppSettings:JWTSecretKey"]!))
+                };
+            });
+
+            return services;
+        }
+
+        public static WebApplication AddIdentityAuthMiddleware(this WebApplication app)
+        {
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            return app;
+        }
+    }
+}
