+++ EndpointX/Extensions/SwaggerExtensions.cs
@@ -1,0 +1,26 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace EndpointX.Extensions
+{
+    public static class SwaggerExtensions
+    {
+        public static IServiceCollection AddSwaggerExplorer(this IServiceCollection services)
+        {
+            services.AddEndpointsApiExplorer();
+            services.AddSwaggerGen();
+
+            return services;
+        }
+
+        public static WebApplication ConfigureSwaggerExplorer(this WebApplication app)
+        {
+            if (app.Environment.IsDevelopment())
+            {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
+            return app;
+        }
+    }
+}
