+++ EndpointX/Models/AppSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace EndpointX.Models
+{
+    public class AppSettings
+    {
+        public string JWTSecretKey { get; set; }
+    }
+}
